+++ cdrs-tokio/src/frame/frame_auth_response.rs
@@ -28,13 +28,12 @@ impl Frame {
     /// Creates new frame of type `AuthResponse`.
     pub fn new_req_auth_response(token_bytes: CBytes) -> Frame {
         let version = Version::Request;
-        let flag = Flag::Ignore;
         let opcode = Opcode::AuthResponse;
         let body = BodyReqAuthResponse::new(token_bytes);
 
         Frame::new(
             version,
-            vec![flag],
+            Flags::empty(),
             opcode,
             body.serialize_to_vec(),
             None,
@@ -61,7 +60,6 @@ mod tests {
         let frame = Frame::new_req_auth_response(CBytes::new(bytes));
 
         assert_eq!(frame.version, Version::Request);
-        assert_eq!(frame.flags, vec![Flag::Ignore]);
         assert_eq!(frame.opcode, Opcode::AuthResponse);
         assert_eq!(frame.body, &[0, 0, 0, 3, 1, 2, 3]);
         assert_eq!(frame.tracing_id, None);